+++ src/docs/HISTORY.md
@@ -11,6 +11,8 @@ v1.4.0
 - Added a Preference to set a multiplier on how slow with Alt + LMB.
 - Added a Preference to control how scrubbing looping behavior works.  It can be set to Button (the UI timeline setting), Inactive or Active.
 - Fixed Vulkan's DPI and scaling of images/videos within the viewport.
+- Made audio stop playback when the window is minimized (due to a bug in FLTK
+  only on Windows).
 
 
 v1.3.9